// FE/src/types/StatusEnum.ts
export enum ApplyStatus {
  REJECTED = "REJECTED",
  APPROVED = "APPROVED",
  PENDING = "PENDING",
  NOT_REGISTER = "NOT_REGISTER"
}

export enum Status {
  ACTIVE = "Active",
  INACTIVE = "Inactive",
  BANNED = "Banned",

}

export enum Role {
  ADMIN = "ADMIN",
  STUDENT = "STUDENT",
  HR_STAFF = "COMPANY",
  STAFF = "StudentServicesDepartmentStaff"
}